#689: provide support for EPS files
EPS is added to the front of the registry to ensure that very small
EPS files will be detected.

This implementation refuses to deal with "read-only XMP" as well
as "%%Include..." DSC comments, because it is unclear how to handle
those properly. If one of these special cases occur, a warning will
be emitted which encourages the user to provide a real-world EPS
file to us.

git-svn-id: svn://dev.exiv2.org/svn/trunk@2479 b7c8b350-86e7-0310-a4b4-de8f6a8f16a3